Extract frontend validator to allow individual import
