+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "satpam",
-  "version": "4.13.1",
+  "version": "4.14.1",
   "description": "Simple and Effective Object Validator",
   "main": "lib/index.js",
   "scripts": {
